requirements.txt
Flask
bs4
joblib
selenium
flask-bootstrap
pillow
